requirements.txt
pysrt
webvtt-py
pysubs2
pycaption
moviepy
torchaudio
torch
silero-vad
pyaudioanalysis

ffmpeg
